@@ -4,7 +4,10 @@
   "private": true,
   "dependencies": {
     "@brainhubeu/react-file-input": "^0.9.20",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
     "@material-ui/core": "^4.12.3",
+    "@mui/material": "^5.0.4",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -18,21 +21,30 @@
     "axios": "^0.21.1",
     "bootstrap": "^4.5.2",
     "browserslist": "^4.14.2",
+    "draft-js": "^0.11.7",
+    "draftjs-to-html": "^0.9.1",
+    "final-form-arrays": "^3.0.2",
     "flutterwave-react-v3": "^1.0.9",
     "font-awesome": "^4.7.0",
     "google-map-react": "^2.1.9",
+    "hoist-non-react-statics": "^3.3.2",
     "moment": "^2.28.0",
-    "react": "^16.13.1",
+    "multiselect-react-dropdown": "^2.0.25",
+    "react": "^17.0.2",
     "react-accessible-accordion": "^3.3.5",
+    "react-animated-text-builders": "^1.1.0",
     "react-apexcharts": "^1.3.9",
     "react-bar-chart": "^0.1.2",
     "react-bootstrap": "^1.3.0",
     "react-bootstrap-range-slider": "^0.3.8",
     "react-calendar": "^3.1.0",
     "react-cv": "^1.2.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
+    "react-draft-wysiwyg": "^1.15.0",
     "react-facebook-auth": "^1.4.0",
     "react-facebook-pixel": "^1.0.4",
+    "react-final-form": "^6.5.9",
+    "react-final-form-arrays": "^3.1.4",
     "react-ga": "^3.3.0",
     "react-google-charts": "^3.0.15",
     "react-google-login": "^5.1.21",
@@ -46,10 +58,11 @@
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-router-hash-link": "^2.4.3",
-    "react-scripts": "3.3.0",
+    "react-scripts": "^5.0.0",
     "react-simple-charts": "0.0.19",
     "react-simple-sidenav": "^1.0.1",
     "react-star-rating-component": "^1.4.1",
+    "react-text-transition": "^1.3.0",
     "react-to-pdf": "0.0.13",
     "react-toastify": "^5.5.0",
     "redux": "^4.0.5",
@@ -57,7 +70,7 @@
     "salad-ui.chart": "^1.1.55",
     "serve": "^11.3.2",
     "styled-components": "^5.3.0",
-    "sweetalert2": "^11.0.17",
+    "sweetalert2": "^11.3.10",
     "sweetalert2-react-content": "^4.0.1",
     "typescript": "^3.9.7"
   },
