@@ -36,6 +36,7 @@ const WorkDetails_Form_Preview = (props) => {
     state_: "",
     specialist_config: [],
     pipe_config: [],
+    spreads:[],
   });
   const {
     pipe_wieght,
@@ -62,6 +63,7 @@ const WorkDetails_Form_Preview = (props) => {
     isloading,
     order_title,
     project_location,
+    spreads
   } = state;
 
   useEffect(() => {
@@ -71,11 +73,13 @@ const WorkDetails_Form_Preview = (props) => {
     console.log(firstData);
     const secondList: any = localStorage.getItem("admin_second_step");
     const secondData = secondList ? JSON.parse(secondList) : "";
-    console.log(secondData);
+    const third_data = localStorage.getItem("spreads")
+    const spreads = third_data?JSON.parse(third_data):""
     setState({
       ...state,
       ...secondData,
       ...firstData,
+      ...spreads,
     });
   }, []);
   const submitForm = () => {
@@ -123,6 +127,7 @@ const WorkDetails_Form_Preview = (props) => {
       pipe_configs: pipe_keys,
       skills: config_keys,
       project_location,
+      spreads,
     };
     console.log(work_order_data);
     axios
@@ -282,6 +287,32 @@ const WorkDetails_Form_Preview = (props) => {
                       </h6>
                       <div className="Construction12">
                         {data?.pipe_schedule ?? "n/a"}
+                      </div>
+                    </div>
+                  </div>
+                </Col>
+              ))}
+               <h6 className="userprofile12 userprofile123 userprofile1231">
+                Spread
+              </h6>
+              {spreads?.map((data, i) => (
+                <Col md={12} className="ttp_" key={i}>
+                  <div className="closticon"></div>
+                  <div className="main_wrap_ws main_wrap_ws22 graybg">
+                    <div>
+                      <h6 className="userprofile12 userprofile123">
+                       Number of Welders
+                      </h6>
+                      <div className="Construction12">
+                        {data?.welders}
+                      </div>
+                    </div>
+                    <div className="">
+                      <h6 className="userprofile12 userprofile123">
+                        Number of Fitters
+                      </h6>
+                      <div className="Construction12">
+                        {data?.fitters}
                       </div>
                     </div>
                   </div>
