@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { Col, Row, Container, Modal, Form,Pagination } from "react-bootstrap";
+import { Col, Row, Container, Modal, Form, Pagination } from "react-bootstrap";
 import "./contractor.css";
 import DashboardNav from "./navbar";
 import portfolio from "../../images/portfolio.png";
@@ -7,18 +7,17 @@ import group2 from "../../images/group2.png";
 import Slider from "react-rangeslider";
 import "react-rangeslider/lib/index.css";
 import { Helmet } from "react-helmet";
-import nextbtn from "../../images/nextbtn.png";
 import WorkOrderCards from "./WorkOrderCards";
 import { Link } from "react-router-dom";
 import no_work_order from "../../images/document 1.png";
-import no_work_order2 from "../../images/calendar 1.png";
+import email3 from "../../images/email3.png";
 import Swal from "sweetalert2";
 import withReactContent from "sweetalert2-react-content";
 import axios from "axios";
-import { API } from "../../config";
-
+import { API, loggedInDetails } from "../../config";
 
 const MySwal = withReactContent(Swal);
+
 const ListOfContractor = () => {
   const [state, setState] = useState({
     contractor_list: [],
@@ -40,14 +39,15 @@ const ListOfContractor = () => {
     rc_number: "",
     company_phone: "",
     company_email: "",
-    next:"",
-    prev:"",
-    first:"",
-    last:"",
-    current_page:"",
-    last_page:"",
-    to:"",
-    total:"",
+    contractor_id: "",
+    next: "",
+    prev: "",
+    first: "",
+    last: "",
+    current_page: "",
+    last_page: "",
+    to: "",
+    total: "",
   });
   const openModal = (e, x) => {
     setState({
@@ -109,14 +109,6 @@ const ListOfContractor = () => {
     }
   };
   const submit = () => {
-    // MySwal.fire({
-    //   title: "Are you sure?",
-    //   text: "You will not be able to recover this imaginary file!",
-    //   icon: "warning",
-    //   showCancelButton: true,
-    //   confirmButtonText: "Yes, delete it!",
-    //   cancelButtonText: "No, keep it",
-    // });
     MySwal.fire({
       title: "Successfully added contractor",
     });
@@ -125,7 +117,7 @@ const ListOfContractor = () => {
       show: false,
     });
   };
-  const nextPage=(x)=>{
+  const nextPage = (x) => {
     const availableToken: any = localStorage.getItem("loggedInDetails");
     const token = availableToken
       ? JSON.parse(availableToken)
@@ -144,40 +136,17 @@ const ListOfContractor = () => {
             ...state,
             contractor_list: res.data.data.data,
             ...res.data.data.links,
-            ...res.data.data.meta
+            ...res.data.data.meta,
           });
         })
       )
       .catch((err) => {
         console.log(err);
       });
-  }
+  };
   useEffect(() => {
     window.scrollTo(-0, -0);
-    const availableToken: any = localStorage.getItem("loggedInDetails");
-    const token = availableToken
-      ? JSON.parse(availableToken)
-      : window.location.assign("/");
-    axios
-      .all([
-        axios.get(`${API}/admin/contractors?paginate=1`, {
-          headers: { Authorization: `Bearer ${token.access_token}` },
-        }),
-      ])
-      .then(
-        axios.spread((res) => {
-          console.log(res.data.data);
-          setState({
-            ...state,
-            contractor_list: res.data.data.data,
-            ...res.data.data.links,
-            ...res.data.data.meta
-          });
-        })
-      )
-      .catch((err) => {
-        console.log(err);
-      });
+    getAllContractors();
   }, []);
   const {
     inprogress,
@@ -191,7 +160,7 @@ const ListOfContractor = () => {
     work_order_description,
     project_purpose,
     location,
-    end_date,
+    contractor_id,
     show,
     company_phone,
     company_email,
@@ -204,6 +173,106 @@ const ListOfContractor = () => {
     to,
     total,
   } = state;
+  const getAllContractors = () => {
+    axios
+      .all([
+        axios.get(`${API}/admin/contractors?paginate=1`, {
+          headers: {
+            Authorization: `Bearer ${loggedInDetails().access_token}`,
+          },
+        }),
+      ])
+      .then(
+        axios.spread((res) => {
+          window.scrollTo(0, 0);
+          console.log(res.data.data);
+          setState({
+            ...state,
+            contractor_list: res.data.data.data,
+            ...res.data.data.links,
+            ...res.data.data.meta,
+            inprogress: true,
+            pending_request: false,
+            past: false,
+          });
+        })
+      )
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  const deleteContractor = (id) => {
+    setState({
+      ...state,
+      contractor_id: id,
+    });
+    Swal.fire({
+      title: "Confirm Delete!",
+      icon: "warning",
+      text: "Everything associated with this contractor including work orders, payment invoices, etc will be completely erased and this action cannot be undone, do you want to continue",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, delete!",
+      cancelButtonText: "No, keep it",
+      preConfirm: () => delete_contractor(id),
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Deleted!", "Contractor has been deleted.", "success");
+      }
+    });
+  };
+
+  const delete_contractor = (id) => {
+    axios
+      .delete(`${API}/admin/contractors/${id}`, {
+        headers: { Authorization: `Bearer ${loggedInDetails().access_token}` },
+      })
+      .then((res) => {
+        console.log(res);
+        getAllContractors();
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  const archive_contractor = (id) => {
+    axios
+      .delete(`${API}/admin/contractors/${id}/archive`, {
+        headers: { Authorization: `Bearer ${loggedInDetails().access_token}` },
+      })
+      .then((res) => {
+        console.log(res);
+        getAllContractors();
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  const archiveContractor = (id) => {
+    setState({
+      ...state,
+      contractor_id: id,
+    });
+    Swal.fire({
+      title: "Confirm Archive!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, archive!",
+      cancelButtonText: "No, back",
+      preConfirm: () => archive_contractor(id),
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Successful!", "Added to archive.", "success");
+      }
+    });
+  };
+  console.log(contractor_id);
   return (
     <>
       <Container fluid={true} className="dasbwr">
@@ -221,31 +290,25 @@ const ListOfContractor = () => {
               <div className="jobs">Contractor</div>
               <div className="job2">
                 <Link to="/contractor_onboarding" target="blank">
-                  <div className="job3">Add Contractor</div>
+                  <div className="job3 jwidth">Add Contractor</div>
                 </Link>
               </div>
             </div>
             <div className="intab">
               <div
-                onClick={() => switchTab("firsttab")}
+                onClick={() => getAllContractors()}
                 className={inprogress ? "inprogress tab_active" : "inprogress"}
               >
                 All
               </div>
-              <div
+              {/* <div
                 onClick={() => switchTab("secondtab")}
                 className={
                   pending_request ? "inprogress tab_active" : "inprogress"
                 }
               >
-                Active
-              </div>
-              <div
-                onClick={() => switchTab("thirdtab")}
-                className={past ? "inprogress tab_active" : "inprogress"}
-              >
-                Inactive
-              </div>
+                Archive
+              </div> */}
             </div>
             <Row>
               {contractor_list?.length == 0 && (
@@ -261,19 +324,32 @@ const ListOfContractor = () => {
                 </Col>
               )}
               <div className="cardflex_jo">
-                {contractor_list?.map((data, i) => (
-                  <WorkOrderCards
-                    contractor={data}
-                    key={i}
-                    title="Nigerian National Petroleum Corporation"
-                  />
-                  // <WorkOrderCards
-                  //   title={"Pipeline construction with Sulejah"}
-                  // />
-                  // <WorkOrderCards
-                  //   title={"Pipeline construction with Sulejah"}
-                  //   status={"Awaiting Approval"}
-                  // />
+                {contractor_list?.map((data: any, i) => (
+                  <>
+                    <div className="text-right">
+                      <button
+                        type="button"
+                        onClick={() => archiveContractor(data?.id)}
+                        className="btn btn-info Deletebtn archiv2"
+                      >
+                        Archive
+                      </button>
+                      {/* <button
+                        type="button"
+                        onClick={() => deleteContractor(data?.id)}
+                        className="btn btn-danger Deletebtn"
+                      >
+                        Delete
+                      </button> */}
+                    </div>
+                    <Link to={`/admin/contractors/${data.id}`}>
+                      <WorkOrderCards
+                        contractor={data}
+                        key={i}
+                        title="Nigerian National Petroleum Corporation"
+                      />
+                    </Link>
+                  </>
                 ))}
               </div>
               <div className="active_member2">
@@ -281,10 +357,10 @@ const ListOfContractor = () => {
                   Displaying <b>{current_page}</b> of <b>{last_page}</b>
                 </div>
                 <Pagination>
-                  <Pagination.First onClick={()=>nextPage(first)}/>
-                  <Pagination.Prev onClick={()=>nextPage(prev)}/>
-                  <Pagination.Next onClick={()=>nextPage(next)}/>
-                  <Pagination.Last  onClick={()=>nextPage(last)}/>
+                  <Pagination.First onClick={() => nextPage(first)} />
+                  <Pagination.Prev onClick={() => nextPage(prev)} />
+                  <Pagination.Next onClick={() => nextPage(next)} />
+                  <Pagination.Last onClick={() => nextPage(last)} />
                 </Pagination>
               </div>
             </Row>
