@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { Col, Row, Container, Form, Modal } from "react-bootstrap";
 import "../contractor.css";
 import DashboardNav from "../navbar";
@@ -7,11 +7,11 @@ import "react-rangeslider/lib/index.css";
 import { Helmet } from "react-helmet";
 import { Link } from "react-router-dom";
 import axios from "axios";
-import { API } from "../../../config";
+import { API, returnAdminToken, notify } from "../../../config";
 
 // import logo from "../../images/Molecular.png";
 
-const Contractor_Profile = () => {
+const AdminProfile = (props) => {
   const [state, setState] = useState({
     work_orders: [],
     country: "",
@@ -19,23 +19,58 @@ const Contractor_Profile = () => {
     secondtab: false,
     thirdtab: false,
     fourthtab: false,
-    show: true,
+    show: false,
+    contractor: {},
     company_name: "",
-    company_url: "",
+    contacts: [],
+    roles: [],
+    logo: "",
+    website_url: "",
     sector: "",
     state_: "",
     address: "",
+    role: "",
     phone_number: "",
     email: "",
     first_name: "",
     last_name: "",
+    role1: "",
+    phone_number1: "",
+    email1: "",
+    first_name1: "",
+    last_name1: "",
     middle_name: "",
+    reason: "",
+    industry: "",
+    old_password: "",
     current_password: "",
     new_password: "",
     confirm_password: "",
-    reason: "",
+    list_of_industries: [],
+    industry_id: "",
+    role2: "",
+    role_id1: "",
+    role_id2: "",
+    role_id3: "",
+    phone_number2: "",
+    email2: "",
+    first_name2: "",
+    last_name2: "",
+    role3: "",
+    phone_number3: "",
+    email3: "",
+    first_name3: "",
+    last_name3: "",
+    isloading: false,
   });
   const onchange = (e) => {
+    if (e.target.id == "industry") {
+      return setState({
+        ...state,
+        industry_id: e.target.value,
+        industry: e.target.value,
+      });
+    }
     console.log(e.target.value);
     setState({
       ...state,
@@ -104,19 +139,205 @@ const Contractor_Profile = () => {
   const SubmitProfile = () => {
     const data = {
       company_name,
-      company_url,
-      sector,
+      // website_url,
+      industry: industry_id,
       country,
-      state_,
+      state: state_,
       address,
     };
+    console.log(data);
+    setState({
+      ...state,
+      isloading: true,
+    });
     axios
-      .post(`${API}/pi`, data)
+      .put(`${API}/contractor`, data, {
+        headers: { Authorization: `Bearer ${returnAdminToken().access_token}` },
+      })
       .then((res) => {
         console.log(res);
+        notify("Update successful");
+        setState({
+          ...state,
+          isloading: false,
+        });
       })
       .catch((err) => {
+        notify("Update failed", "D");
+        setState({
+          ...state,
+          isloading: false,
+        });
+      });
+  };
+  const SubmitPassword = () => {
+    const data = {
+      old_password: current_password,
+      password: new_password,
+      password_confirmation: confirm_password,
+    };
+    console.log(data);
+    setState({
+      ...state,
+      isloading: true,
+    });
+    axios
+      .post(`${API}/password`, data, {
+        headers: { Authorization: `Bearer ${returnAdminToken().access_token}` },
+      })
+      .then((res) => {
+        console.log(res);
+        notify("Update successful");
+        setState({
+          ...state,
+          isloading: false,
+        });
+      })
+      .catch((err) => {
+        notify("Update failed", "D");
+        setState({
+          ...state,
+          isloading: false,
+        });
+      });
+  };
+  useEffect(() => {
+    window.scrollTo(0, 0);
+    const token = returnAdminToken();
+    axios
+      .all([
+        axios.get(`${API}/admin`, {
+          headers: { Authorization: `Bearer ${token.access_token}` },
+        }),
+        axios.get(`${API}/industries`, {
+          headers: { Authorization: `Bearer ${token.access_token}` },
+        }),
+        axios.get(`${API}/contractor/contact/roles`, {
+          headers: { Authorization: `Bearer ${token.access_token}` },
+        }),
+      ])
+      .then(
+        axios.spread((res, res2, res3) => {
+          console.log(res.data);
+          setState({
+            ...state,
+            ...res.data.data,
+            list_of_industries: res2.data.data,
+            roles: res3.data.data,
+            contractor: res.data.data,
+            state_: res.data.data.state,
+            industry_id: res.data.data.industry_id,
+            role1: res?.data?.data?.contacts?.[0]?.role,
+            phone_number1: res?.data?.data?.contacts?.[0]?.phone,
+            email1: res?.data?.data?.contacts?.[0]?.email,
+            first_name1: res?.data?.data?.contacts?.[0]?.first_name,
+            last_name1: res?.data?.data?.contacts?.[0]?.last_name,
+            role2: res?.data?.data?.contacts?.[1]?.role,
+            phone_number2: res?.data?.data?.contacts?.[1]?.phone,
+            email2: res?.data?.data?.contacts?.[1]?.email,
+            first_name2: res?.data?.data?.contacts?.[1]?.first_name,
+            last_name2: res?.data?.data?.contacts?.[1]?.last_name,
+            role3: res?.data?.data?.contacts?.[2]?.role,
+            phone_number3: res?.data?.data?.contacts?.[2]?.phone,
+            email3: res?.data?.data?.contacts?.[2]?.email,
+            first_name3: res?.data?.data?.contacts?.[2]?.first_name,
+            last_name3: res?.data?.data?.contacts?.[2]?.last_name,
+            role_id1: res?.data?.data?.contacts?.[0]?.role_id,
+            role_id2: res?.data?.data?.contacts?.[1]?.role_id,
+            role_id3: res?.data?.data?.contacts?.[2]?.role_id,
+          });
+        })
+      )
+      .catch((err) => {
         console.log(err);
+      });
+  }, []);
+  const hiddenFileInput: any = useRef();
+  const handleClick = (event) => {
+    hiddenFileInput.current.click();
+  };
+  const handleImageChange = (e) => {
+    const reader: any = new FileReader();
+    reader.onload = () => {
+      if (reader.readyState === 2) {
+        setState({
+          ...state,
+          logo: reader.result,
+        });
+      }
+    };
+    reader.readAsDataURL(e.target.files[0]);
+    console.log(logo);
+    console.log(e.target.files[0]);
+    // upload image to server;
+    const availableToken = localStorage.getItem("loggedInDetails");
+    console.log(availableToken);
+    const token = availableToken ? JSON.parse(availableToken) : "";
+    console.log(token);
+    const imageData = new FormData();
+    imageData.append("logo", e.target.files[0]);
+    console.log(imageData);
+    axios
+      .post(`${API}/contractor/logo`, imageData, {
+        headers: {
+          Authorization: `Bearer ${token.access_token}`,
+          Accept: "application/json",
+          "Content-Type": "multipart/form-data",
+        },
+      })
+      .then((res) => {
+        console.log(res.data);
+        setTimeout(() => {
+          notify("image uploaded Successfully");
+        }, 1000);
+      })
+      .catch((err) => {
+        console.log(err.response);
+        notify("failed to Upload Image");
+      });
+  };
+  const UpdateContactPerson = (
+    id: any,
+    first_name,
+    last_name,
+    email,
+    phone,
+    role
+  ) => {
+    const data = {
+      first_name,
+      last_name,
+      email,
+      phone,
+      role,
+    };
+    console.log(data);
+    setState({
+      ...state,
+      isloading: true,
+    });
+    axios
+      .put(`${API}/contractor/contacts/${id}`, data, {
+        headers: { Authorization: `Bearer ${returnAdminToken().access_token}` },
+      })
+      .then((res) => {
+        console.log(res);
+        notify("Updated successfully");
+        setState({
+          ...state,
+          isloading: false,
+        });
+      })
+      .catch((err) => {
+        setState({
+          ...state,
+          isloading: false,
+        });
+        if (err?.response?.status == 406) {
+          notify(err?.response?.data?.errors?.email?.join(""), "D");
+          return notify(err?.response?.data?.errors?.role?.join(""), "D");
+        }
+        notify("Failed to update", "D");
       });
   };
   const {
@@ -126,21 +347,47 @@ const Contractor_Profile = () => {
     fourthtab,
     show,
     company_name,
-    company_url,
-    sector,
+    website_url,
+    contacts,
+    role,
+    logo,
     country,
+    roles,
+    old_password,
     state_,
     address,
     phone_number,
     first_name,
     last_name,
-    middle_name,
+    industry,
+    industry_id,
+    isloading,
+    list_of_industries,
     email,
     reason,
     current_password,
     new_password,
     confirm_password,
-  } = state;
+    role1,
+    phone_number1,
+    email1,
+    first_name1,
+    last_name1,
+    role2,
+    phone_number2,
+    email2,
+    first_name2,
+    last_name2,
+    role3,
+    phone_number3,
+    email3,
+    first_name3,
+    last_name3,
+    role_id1,
+    role_id2,
+    role_id3,
+  }: any = state;
+  console.log(contacts);
   return (
     <>
       <Container fluid={true} className="dasbwr">
@@ -160,22 +407,6 @@ const Contractor_Profile = () => {
             <div className="psettings">
               <div className="settings11">
                 <div className="titleprofile1">Settings</div>
-                <div className="setting1">
-                  <div className="namestyle11">
-                    <span className="namestyle">HH</span>
-                    <div className="home_pone12">
-                      <div className="helmot">Helmotz Holdings</div>
-                      <div className="helmot11">Oil and Gas</div>
-                      <div className="helmot112">Lagos State, Nigeria</div>
-                    </div>
-                  </div>
-                  <div className="orders1">
-                    Total Work Orders
-                    <div>
-                      <span className="num12a">2</span>
-                    </div>
-                  </div>
-                </div>
                 <div className="section_form">
                   <div className="profile__001">
                     {firsttab ? (
@@ -301,7 +532,7 @@ const Contractor_Profile = () => {
                           </Form.Group>
                         </Col>
                         <Col md={4} className="formsection1">
-                          <Form.Group>
+                          {/* <Form.Group>
                             <h6 className="userprofile userprofile12">
                               Sector
                             </h6>
@@ -313,10 +544,10 @@ const Contractor_Profile = () => {
                               onChange={onchange}
                               placeholder="yyyy-mm-dd"
                             />
-                          </Form.Group>
+                          </Form.Group> */}
                         </Col>
                         <Col md={4} className="formsection1">
-                          <Form.Group>
+                          {/* <Form.Group>
                             <h6 className="userprofile userprofile12">
                               Company URL
                             </h6>
@@ -327,7 +558,7 @@ const Contractor_Profile = () => {
                               value={company_url}
                               onChange={onInputChange}
                             />
-                          </Form.Group>
+                          </Form.Group> */}
                         </Col>
                       </Row>
                     </>
@@ -626,8 +857,8 @@ const Contractor_Profile = () => {
           <Row>
             <Col md={12}>
               <p>
-                You are about to deactivate this account, please note that this action cannot
-                be undone.
+                You are about to deactivate this account, please note that this
+                action cannot be undone.
               </p>
             </Col>
           </Row>
@@ -642,4 +873,4 @@ const Contractor_Profile = () => {
   );
 };
 
-export default Contractor_Profile;
+export default AdminProfile;
