@@ -46,18 +46,17 @@ const Play = () => {
   return (
     <>
       <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="20"
-        fill="white"
-        viewBox="0 0 26 26"
-      >
+        xmlns='http://www.w3.org/2000/svg'
+        width='20'
+        fill='white'
+        viewBox='0 0 26 26'>
         <polygon
-          className="play-btn__svg"
-          points="9.33 6.69 9.33 19.39 19.3 13.04 9.33 6.69"
+          className='play-btn__svg'
+          points='9.33 6.69 9.33 19.39 19.3 13.04 9.33 6.69'
         />
         <path
-          className="play-btn__svg"
-          d="M26,13A13,13,0,1,1,13,0,13,13,0,0,1,26,13ZM13,2.18A10.89,10.89,0,1,0,23.84,13.06,10.89,10.89,0,0,0,13,2.18Z"
+          className='play-btn__svg'
+          d='M26,13A13,13,0,1,1,13,0,13,13,0,0,1,26,13ZM13,2.18A10.89,10.89,0,1,0,23.84,13.06,10.89,10.89,0,0,0,13,2.18Z'
         />
       </svg>
     </>
@@ -66,7 +65,7 @@ const Play = () => {
 const Stop = () => {
   return (
     <>
-      <div className="stop3"></div>
+      <div className='stop3'></div>
     </>
   );
 };
@@ -102,13 +101,13 @@ const Work_Details = (props: any) => {
     <>
       <Card>
         <div onClick={toggleAccordion1}>
-          <div className="deploydsplstwrapp">
+          <div className='deploydsplstwrapp'>
             <div>
-              <span className="deplyeaggrgt">
+              <span className='deplyeaggrgt'>
                 {props?.group_data?.total_members} Work Details
               </span>
             </div>
-            <div className="accimgwrap">
+            <div className='accimgwrap'>
               <span>
                 <img src={chevrondown} className={`arrow-down ${chevron}`} />
               </span>
@@ -116,12 +115,11 @@ const Work_Details = (props: any) => {
           </div>
         </div>
       </Card>
-      {isloading && <Spinner animation={"grow"} variant="info" />}
+      {isloading && <Spinner animation={"grow"} variant='info' />}
       <div
         style={{ maxHeight: `${collapseHeight}` }}
-        className="acccollapsediv1"
-        ref={content1}
-      >
+        className='acccollapsediv1'
+        ref={content1}>
         <>
           <WorkDetails_Form_Preview
             order_detail={props.work_order_detailz}
@@ -212,11 +210,11 @@ const Work_Sheet = (props: any) => {
     <>
       <Card>
         <div onClick={toggleAccordion1}>
-          <div className="deploydsplstwrapp">
+          <div className='deploydsplstwrapp'>
             <div>
-              <span className="deplyeaggrgt">Work Sheets</span>
+              <span className='deplyeaggrgt'>Work Sheets</span>
             </div>
-            <div className="accimgwrap">
+            <div className='accimgwrap'>
               <span>
                 <img src={chevrondown} className={`arrow-down ${chevron}`} />
               </span>
@@ -224,43 +222,43 @@ const Work_Sheet = (props: any) => {
           </div>
         </div>
       </Card>
-      {isloading && <Spinner animation={"grow"} variant="info" />}
+      {isloading && <Spinner animation={"grow"} variant='info' />}
       <div
         style={{ maxHeight: `${collapseHeight}` }}
-        className="acccollapsediv1"
-        ref={content1}
-      >
+        className='acccollapsediv1'
+        ref={content1}>
         <>
-          <div className="worksheet_1">
+          <div className='worksheet_1'>
             {work_sheet.map((data: any, i) => (
-              <div className="tabledata tablecontent tablecont1">
-                <div className="header_12 tablecont0">
+              <div className='tabledata tablecontent tablecont1'>
+                <div className='header_12 tablecont0'>
                   <span>Worksheet Report {data.week}</span>
                 </div>
-                <div className="tablecont1">
-                  <div className="worksheetdw worksheetdate1">
+                <div className='tablecont1'>
+                  <div className='worksheetdw worksheetdate1'>
                     {" "}
-                    <img src={dwnload} alt="dwnload" className="dwnload1" />
+                    <img src={dwnload} alt='dwnload' className='dwnload1' />
                     <a href={data.worksheet} target={"blank"}>
                       Download
                     </a>
                   </div>
-                  <div className="worksheetdate">{formatTime(data.date)}</div>
-                  <div className="upby">uploaded by {data.uploaded_by}</div>
-                  <div className="upby">
+                  <div className='worksheetdate'>{formatTime(data.date)}</div>
+                  <div className='upby'>
+                    uploaded by <br /> {data.uploaded_by}
+                  </div>
+                  <div className='upby'>
                     {" "}
                     <div>
                       {data.approved ? "Approved" : "Awaiting Approval"}
                     </div>
                   </div>
-                  <div className="upby">
+                  <div className='upby'>
                     {data.sent ? (
                       ""
                     ) : (
                       <span
-                        className="raise_inv"
-                        onClick={() => sendToContractor(data.id)}
-                      >
+                        className='raise_inv'
+                        onClick={() => sendToContractor(data.id)}>
                         {isloading ? "Sending" : "Send"}
                       </span>
                     )}
@@ -269,15 +267,15 @@ const Work_Sheet = (props: any) => {
               </div>
             ))}
             {work_sheet?.length == 0 && (
-              <Col md={11} className="containerforemptyorder1 cust20">
-                <div className="containerforemptyorder">
+              <Col md={11} className='containerforemptyorder1 cust20'>
+                <div className='containerforemptyorder'>
                   <img
                     src={no_work_order}
                     alt={"no_work_order"}
-                    className="no_work_order"
+                    className='no_work_order'
                   />
                 </div>
-                <div className="no_work1">worksheet has not been sent</div>
+                <div className='no_work1'>worksheet has not been sent</div>
               </Col>
             )}
           </div>
@@ -295,8 +293,8 @@ const Upfront_payment = (props: any) => {
     upfront: [],
     work_id: "",
     reason: "",
-    show:false,
-    id:""
+    show: false,
+    id: "",
   });
   const content1: any = useRef();
   const toggleAccordion1 = () => {
@@ -376,10 +374,10 @@ const Upfront_payment = (props: any) => {
     setState({
       ...state,
       show: true,
-      id
+      id,
     });
   };
- 
+
   const declineUpfrontRequest = () => {
     setState({
       ...state,
@@ -425,12 +423,12 @@ const Upfront_payment = (props: any) => {
     work_id,
     reason,
     show,
-    id
+    id,
   } = state;
   return (
     <>
-    <Modal
-        size="lg"
+      <Modal
+        size='lg'
         show={show}
         onHide={() =>
           setState({
@@ -438,12 +436,11 @@ const Upfront_payment = (props: any) => {
             show: false,
           })
         }
-        dialogClassName="modal-90w"
-        className="mdl12"
-      >
+        dialogClassName='modal-90w'
+        className='mdl12'>
         <Modal.Header closeButton>
-          <Modal.Title id="example-custom-modal-styling-title">
-            Reject 
+          <Modal.Title id='example-custom-modal-styling-title'>
+            Reject
           </Modal.Title>
         </Modal.Header>
         <Modal.Body>
@@ -454,28 +451,30 @@ const Upfront_payment = (props: any) => {
                   value={reason}
                   name={"reason"}
                   onChange={onchange}
-                  className="form-control reason12 reason122"
-                  placeholder="Please leave a message"
-                ></textarea>
+                  className='form-control reason12 reason122'
+                  placeholder='Please leave a message'></textarea>
               </Form>
             </Col>
           </Row>
           <Row>
-            <Col md={12} className="terminate2">
-              <div className="terminate1" onClick={(e) => declineUpfrontRequest()}>
-                {isloading?"Rejecting":"Reject"}
+            <Col md={12} className='terminate2'>
+              <div
+                className='terminate1'
+                onClick={(e) => declineUpfrontRequest()}>
+                {isloading ? "Rejecting" : "Reject"}
               </div>
             </Col>
           </Row>
         </Modal.Body>
       </Modal>
+
       <Card>
         <div onClick={toggleAccordion1}>
-          <div className="deploydsplstwrapp">
+          <div className='deploydsplstwrapp'>
             <div>
-              <span className="deplyeaggrgt">Early Payment Request</span>
+              <span className='deplyeaggrgt'>Early Payment Request</span>
             </div>
-            <div className="accimgwrap">
+            <div className='accimgwrap'>
               <span>
                 <img src={chevrondown} className={`arrow-down1 ${chevron}`} />
               </span>
@@ -483,64 +482,62 @@ const Upfront_payment = (props: any) => {
           </div>
         </div>
       </Card>
-      {isloading && <Spinner animation={"grow"} variant="info" />}
+      {isloading && <Spinner animation={"grow"} variant='info' />}
       <div
         style={{ maxHeight: `${collapseHeight}` }}
-        className="acccollapsediv1"
-        ref={content1}
-      >
+        className='acccollapsediv1'
+        ref={content1}>
         <>
-          <div className="worksheet_1">
+          <div className='worksheet_1'>
             {upfront.map((data: any, i) => (
-              <div className="tabledata tablecontent tablecont1">
-                <div className="header_12 tablecont0">
+              <div className='tabledata tablecontent tablecont1'>
+                <div className='header_12 tablecont0'>
                   <span>{data.specialist}</span>
-                  <div className="lightgray">{data.created}</div>
+                  <div className='lightgray'>{data.created}</div>
                 </div>
-                <div className="tablecont1">
-                  <div className="worksheetdw worksheetdate1"> </div>
-                  <div className="upby awaiting9">{current_currency}{FormatAmount(data.amount)}</div>
-                  <div className="upby awaiting9">
-                    {" "}
-                    <div>
-                      {data.approved ? "Approved" : ""}
-                    </div>
+                <div className='tablecont1'>
+                  <div className='worksheetdw worksheetdate1'> </div>
+                  <div className='upby awaiting9'>
+                    {current_currency}
+                    {FormatAmount(data.amount)}
                   </div>
-                  <div className="upby accrjct">
-                    {data.status=="Paid" ? (
+                  <div className='upby awaiting9'>
+                    {" "}
+                    <div>{data.approved ? "Approved" : ""}</div>
+                  </div>
+                  <div className='upby accrjct'>
+                    {data.status == "Paid" ? (
                       "Paid"
-                    ) :
-                      data.status=="Declined"? "Declined":(
-                        <>
-                          <span
-                            className="raise_inv"
-                            onClick={() => approveUpfrontRequest(data.id)}
-                          >
-                            {isloading ? "Approving" : "Approve"}
-                          </span>
-                          <span
-                            className="raise_inv reje4"
-                            onClick={() => openModal(data.id)}
-                          >
-                            {"Decline"}
-                          </span>
-                        </>
-                      )
-                    }
+                    ) : data.status == "Declined" ? (
+                      "Declined"
+                    ) : (
+                      <>
+                        <span
+                          className='raise_inv'
+                          onClick={() => approveUpfrontRequest(data.id)}>
+                          {isloading ? "Approving" : "Approve"}
+                        </span>
+                        <span
+                          className='raise_inv reje4'
+                          onClick={() => openModal(data.id)}>
+                          {"Decline"}
+                        </span>
+                      </>
+                    )}
                   </div>
                 </div>
               </div>
             ))}
             {upfront?.length == 0 && (
-              <Col md={11} className="containerforemptyorder1 cust20">
-                <div className="containerforemptyorder">
+              <Col md={11} className='containerforemptyorder1 cust20'>
+                <div className='containerforemptyorder'>
                   <img
                     src={no_work_order}
                     alt={"no_work_order"}
-                    className="no_work_order"
+                    className='no_work_order'
                   />
                 </div>
-                <div className="no_work1">data is empty</div>
+                <div className='no_work1'>data is empty</div>
               </Col>
             )}
           </div>
@@ -573,11 +570,11 @@ const Invoice_details = ({ work_order_detail }: any) => {
     <>
       <Card>
         <div onClick={toggleAccordion1}>
-          <div className="deploydsplstwrapp">
+          <div className='deploydsplstwrapp'>
             <div>
-              <span className="deplyeaggrgt">Invoice</span>
+              <span className='deplyeaggrgt'>Invoice</span>
             </div>
-            <div className="accimgwrap">
+            <div className='accimgwrap'>
               <span>
                 <img src={chevrondown} className={`arrow-down1 ${chevron}`} />
               </span>
@@ -585,53 +582,51 @@ const Invoice_details = ({ work_order_detail }: any) => {
           </div>
         </div>
       </Card>
-      {isloading && <Spinner animation={"grow"} variant="info" />}
+      {isloading && <Spinner animation={"grow"} variant='info' />}
       <div
         style={{ maxHeight: `${collapseHeight}` }}
-        className="acccollapsediv1"
-        ref={content1}
-      >
+        className='acccollapsediv1'
+        ref={content1}>
         <>
-          <div className="job23_1a wrap_z">
-            <div className="main_wrap_ws main_wrapp1">
-              <h6 className="userprofile12 userprofile123"></h6>
-              <div className="tabledata tablecontent">
-                <div className="header_12">Invoice Number</div>
-                <div className="header_12">Total Amount</div>
-                <div className="header_12">Amount Paid</div>
-                <div className="header_12 ">Outstanding</div>
-                <div className="header_12">Total cycles</div>
+          <div className='job23_1a wrap_z'>
+            <div className='main_wrap_ws main_wrapp1'>
+              <h6 className='userprofile12 userprofile123'></h6>
+              <div className='tabledata tablecontent'>
+                <div className='header_12'>Invoice Number</div>
+                <div className='header_12'>Total Amount</div>
+                <div className='header_12'>Amount Paid</div>
+                <div className='header_12 '>Outstanding</div>
+                <div className='header_12'>Total cycles</div>
               </div>
-              <div className="tabledata tablecontent">
-                <div className="header_12">
+              <div className='tabledata tablecontent'>
+                <div className='header_12'>
                   {work_order_detail?.invoice?.reference}
                 </div>
-                <div className="header_12">
+                <div className='header_12'>
                   {current_currency}
                   {FormatAmount(work_order_detail?.invoice?.total_amount)}
                 </div>
-                <div className="header_12">
+                <div className='header_12'>
                   {current_currency}
                   {FormatAmount(work_order_detail?.invoice?.total_amount_paid)}
                 </div>
-                <div className="header_12 active_member">
+                <div className='header_12 active_member'>
                   {current_currency}
                   {FormatAmount(
                     work_order_detail?.invoice?.total_amount_unpaid
                   )}
                 </div>
-                <div className="header_12 active_member">
+                <div className='header_12 active_member'>
                   {work_order_detail?.invoice?.total_cycles}
                 </div>
               </div>
-              <div className="text-center">
+              <div className='text-center'>
                 {" "}
-                <span className="viewall_">
+                <span className='viewall_'>
                   {" "}
                   <Link
-                    to={`/admin_invoice_details/${work_order_detail.invoice?.id}`}
-                    title="view payment cycle information"
-                  >
+                    to={`/admin_invoice_details/${work_order_detail?.invoice?.id}/${work_order_detail?.invoice?.work_order?.id}`}
+                    title='view payment cycle information'>
                     view more
                   </Link>{" "}
                 </span>{" "}
@@ -658,6 +653,9 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
     location: "",
     end_date: "",
     start_date: "",
+    show2: "",
+    notify_specialist: false,
+    specialist_id: "",
     isloading: false,
     hour: "",
     show: false,
@@ -668,18 +666,38 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
     active: "",
     active1: "",
     collapseHeight: "0px",
+    cost_show: false,
     chevron: "",
     chevron1: "",
     allAssignedSpecialist: [],
     work_sheet: [],
+    markup_percentage: "",
+    health_insurance_cost: "",
+    coverall_cost: "",
+    per_diem: "",
+    project_duration: "",
+    spreads: "",
+    professional_indemnity_insurance: "",
+    cost_per_inch: "",
+    association_cost_per_inch: "",
+    fx_rate: "",
+    cost_per_inch_dollar: "",
+    per_diem_dollar: "",
   });
   const onchange = (e) => {
-    console.log(e.target.value);
     setState({
       ...state,
       [e.target.name]: e.target.value,
     });
+    // if (e.target.name == "cost_per_inch_dollar") {
+    // console.log(calc_cost_per_inc_naira(cost_per_inch_dollar))
+    // }
   };
+
+  const calc_cost_per_inc_naira = (d) => {
+    return d * fx_rate;
+  };
+
   const onInputChange = (e) => {
     const letterNumber = /^[A-Za-z]+$/;
     if (e.target.value) {
@@ -708,12 +726,18 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
   const {
     assigned_specialists,
     country,
-    work_order_description,
+    cost_show,
     order_title,
     end_date,
     already_approved,
+    show2,
+    specialist_id,
+    notify_specialist,
     reason,
     location_terrain,
+    cost_per_inch,
+    association_cost_per_inch,
+    fx_rate,
     new_work,
     start_date,
     work_order_detail,
@@ -722,10 +746,19 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
     active,
     active1,
     collapseHeight,
-    chevron,
+    markup_percentage,
+    health_insurance_cost,
+    coverall_cost,
+    per_diem,
+    project_duration,
+    spreads,
+    professional_indemnity_insurance,
     chevron1,
     isloading,
+    cost_per_inch_dollar,
+    per_diem_dollar,
   } = state;
+
   const Accept_work_order = () => {
     const availableToken: any = localStorage.getItem("loggedInDetails");
     const token = availableToken
@@ -756,7 +789,9 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
             JSON.stringify(props.order_details)
           );
           setTimeout(() => {
-            props.history.push("/admin_work_details?inreview=true");
+            props.history.push(
+              `/admin_work_details${props.order_details.id}?inreview=true`
+            );
           }, 2000);
           console.log(res.data);
           setState({
@@ -787,7 +822,7 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
     const work_order = localStorage.getItem("work_order_details");
     const work_order_details = work_order ? JSON.parse(work_order) : "";
     axios
-      .get(`${API}/admin/work-orders/${work_order_details?.id}`, {
+      .get(`${API}/admin/work-orders/${props.match.params.id}`, {
         headers: { Authorization: `Bearer ${token.access_token}` },
       })
       .then((res) => {
@@ -795,6 +830,18 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
         setState({
           ...state,
           ...res.data.data,
+          ...res.data.data.costing,
+          health_insurance_cost:
+            res?.data?.data?.costing?.health_insurance_per_specialist,
+          coverall_cost: res?.data?.data?.costing?.coverall_per_specialist,
+          per_diem: res?.data?.data?.costing?.per_diem_per_specialist,
+          cost_per_inch_dollar:
+            res?.data?.data?.costing?.cost_per_inch /
+              res?.data?.data?.costing?.fx_rate ?? 0,
+          per_diem_dollar:
+            res?.data?.data?.costing?.per_diem_per_specialist /
+              res?.data?.data?.costing?.fx_rate ?? 0,
+          project_duration: res?.data?.data?.costing?.duration_in_days,
           already_approved: urlkey ? true : false,
           work_order_detail: res.data.data,
           new_work: res.data.data.status == "New" ? true : false,
@@ -907,10 +954,177 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
     });
   };
 
+  const modalShow1 = () => {
+    setState({
+      ...state,
+      cost_show: true,
+    });
+  };
+
+
+
+  const sendSLA = () => {
+    const work_order = localStorage.getItem("work_order_details");
+    const work_order_details = work_order ? JSON.parse(work_order) : "";
+    setState({
+      ...state,
+      isloading: true,
+    });
+    axios
+      .post(
+        `${API}/admin/work-orders/${work_order_details?.id}/send-contract`,
+        {},
+        {
+          headers: {
+            Authorization: `Bearer ${returnAdminToken().access_token}`,
+          },
+        }
+      )
+      .then((res) => {
+        console.log(res.data);
+        notify("successfully  sent contract");
+        setTimeout(() => {
+          // window.location.reload();
+        }, 2000);
+        setState({
+          ...state,
+          isloading: false,
+        });
+      })
+      .catch((err) => {
+        if (err?.response?.status == 400) {
+          notify(err?.response?.data?.message);
+        }
+        notify("failed to process");
+        setState({
+          ...state,
+          isloading: false,
+        });
+        console.log(err);
+      });
+  };
+  const modalShow2 = (id) => {
+    setState({
+      ...state,
+      show2: true,
+      specialist_id: id,
+    });
+  };
+
+  const removeSpecialist = () => {
+    setState({
+      ...state,
+      isloading: true,
+    });
+    const data = {
+      specialists: [specialist_id],
+      notify: notify_specialist,
+    };
+    axios
+      .post(
+        `${API}/admin/work-orders/${props.match.params.id}/unassign-specialists`,
+        data,
+        {
+          headers: {
+            Authorization: `Bearer ${returnAdminToken().access_token}`,
+          },
+        }
+      )
+      .then((res) => {
+        console.log(res.data);
+        notify("successfully  removed specialist");
+        setTimeout(() => {
+          window.location.reload();
+        }, 2000);
+        setState({
+          ...state,
+          isloading: false,
+        });
+      })
+      .catch((err) => {
+        if (err?.response?.status == 400) {
+          notify(err?.response?.data?.message);
+        }
+        notify("failed to process");
+        setState({
+          ...state,
+          isloading: false,
+        });
+        console.log(err);
+      });
+  };
+  const SendCost = () => {
+    const work_order = localStorage.getItem("work_order_details");
+    const work_order_details = work_order ? JSON.parse(work_order) : "";
+    setState({
+      ...state,
+      isloading: true,
+    });
+    const data = {
+      markup_percentage,
+      health_insurance_cost,
+      coverall_cost,
+      per_diem: fx_rate * per_diem_dollar,
+      project_duration,
+      spreads,
+      professional_indemnity_insurance,
+      cost_per_inch: fx_rate * cost_per_inch_dollar,
+      association_cost_per_inch,
+      fx_rate,
+    };
+    axios
+      .post(`${API}/admin/work-orders/${work_order_details?.id}/cost`, data, {
+        headers: {
+          Authorization: `Bearer ${returnAdminToken().access_token}`,
+        },
+      })
+      .then((res) => {
+        console.log(res.data);
+        notify("successfully computed work order cost");
+        setTimeout(() => {
+          window.location.reload();
+        }, 2000);
+        setState({
+          ...state,
+          ...res.data.data,
+          isloading: false,
+        });
+      })
+      .catch((err) => {
+        if (err?.response?.status == 400) {
+          notify(err?.response?.data?.message);
+        }
+        notify("failed to process");
+        setState({
+          ...state,
+          isloading: false,
+        });
+        console.log(err);
+      });
+  };
+
+  const calc_cost_per_inch_in_naira = () => {
+    if (fx_rate && cost_per_inch_dollar) {
+      let num = fx_rate * cost_per_inch_dollar;
+      num.toFixed(2);
+      return FormatAmount(num ?? 0);
+    }
+    return 0;
+  };
+
+  const calc_cost_per_diem_in_naira = () => {
+    if (fx_rate && per_diem_dollar) {
+      let num = fx_rate * per_diem_dollar;
+      num.toFixed(2);
+      return FormatAmount(num ?? 0);
+    }
+    return 0;
+  };
+
   return (
     <>
       <Modal
-        size="lg"
+        size='lg'
         show={show}
         onHide={() =>
           setState({
@@ -918,11 +1132,10 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
             show: false,
           })
         }
-        dialogClassName="modal-90w"
-        className="mdl12"
-      >
+        dialogClassName='modal-90w'
+        className='mdl12'>
         <Modal.Header closeButton>
-          <Modal.Title id="example-custom-modal-styling-title">
+          <Modal.Title id='example-custom-modal-styling-title'>
             Reject order
           </Modal.Title>
         </Modal.Header>
@@ -934,14 +1147,13 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                   value={reason}
                   name={"reason"}
                   onChange={onchange}
-                  className="form-control reason12 reason122"
-                  placeholder="Reason for termination"
-                ></textarea>
+                  className='form-control reason12 reason122'
+                  placeholder='Reason for termination'></textarea>
               </Form>
             </Col>
           </Row>
           <Row>
-            <Col md={12} className="terminate2">
+            <Col md={12} className='terminate2'>
               {/* <div className="terminate1" onClick={(e) => Reject_work_order()}>
                 Reject
               </div> */}
@@ -949,29 +1161,341 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
           </Row>
         </Modal.Body>
       </Modal>
-      <Container fluid={true} className="dasbwr">
+      <Modal
+        size='sm'
+        show={show2}
+        onHide={() =>
+          setState({
+            ...state,
+            show2: false,
+          })
+        }
+        dialogClassName=''>
+        <Modal.Header closeButton>
+          <Modal.Title id='example-custom-modal-styling-title'>
+            Remove Specialist
+          </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <Row>
+            <Col md={12}>
+              <p className='teal-text'>
+                Please confirm removal of invited specialist
+              </p>
+              <p>
+                Notify Specialist{" "}
+                <input
+                  type='checkbox'
+                  name='notify_specialist'
+                  className='form-check-input position-static'
+                  onChange={() => {
+                    setState({
+                      ...state,
+                      notify_specialist: notify_specialist ? false : true,
+                    });
+                  }}
+                  checked={notify_specialist ? true : false}
+                  id='notify_specialist '
+                />{" "}
+              </p>
+            </Col>
+          </Row>
+          <Row>
+            <Col md={12} className='terminate2'>
+              <div className='terminate1' onClick={(e) => removeSpecialist()}>
+                {isloading ? "Processing" : "Submit"}
+              </div>
+            </Col>
+          </Row>
+        </Modal.Body>
+      </Modal>
+      <Modal
+        size='lg'
+        show={cost_show}
+        onHide={() =>
+          setState({
+            ...state,
+            cost_show: false,
+          })
+        }
+        dialogClassName='modal-90w'
+        className='mdl12'>
+        <Modal.Header closeButton>
+          <Modal.Title id='example-custom-modal-styling-title'>
+            Cost Settings
+          </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <Row>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Markup Percentage %<span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='markup_percentage'
+                    value={markup_percentage}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Health Insurance per Specialist (₦)
+                  <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='health_insurance_cost'
+                    value={health_insurance_cost}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+          </Row>
+          <Row>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Coverall per Specialist (₦){" "}
+                  <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='coverall_cost'
+                    value={coverall_cost}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Association Cost Per Inch (₦)
+                  <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    name='association_cost_per_inch'
+                    value={association_cost_per_inch}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+          </Row>
+          <Row>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Project Duration (days) <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='project_duration'
+                    value={project_duration}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Spreads <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    name='spreads'
+                    min={0}
+                    value={spreads}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+          </Row>
+          <Row>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Professional Indemnity Insurance Per Spread (₦)
+                  <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='professional_indemnity_insurance'
+                    value={professional_indemnity_insurance}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  FX RATE (₦) <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder='US Dollar to naira rate eg. 550'
+                    name='fx_rate'
+                    min={0}
+                    value={fx_rate}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+          </Row>
+          <Row>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Cost Per Inch ($)<span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='cost_per_inch_dollar'
+                    value={cost_per_inch_dollar}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Cost Per Inch (₦)<span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='text'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    // name="cost_per_inch"
+                    value={calc_cost_per_inch_in_naira()}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+          </Row>
+          {/* Dollar */}
+          <Row>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Per Diem Per Specialist Per Day ($)
+                  <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='number'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='per_diem_dollar'
+                    value={per_diem_dollar}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+            <Col md={6}>
+              <div className='pipelength pipelng'>
+                <div className='pipelength1q'>
+                  Per Diem Per Specialist Per Day (₦)
+                  <span className='text-danger'>*</span>
+                </div>
+                <div className=''>
+                  <input
+                    type='text'
+                    className=' form-control'
+                    placeholder=''
+                    min={0}
+                    name='per_diem'
+                    value={calc_cost_per_diem_in_naira()}
+                    onChange={onchange}
+                  />
+                </div>
+              </div>
+            </Col>
+          </Row>
+          <Row>
+            <Col md={12} className='terminate2'>
+              <Button
+                className=' raise_inv computecost'
+                onClick={(e) => SendCost()}>
+                {isloading ? "processing" : "Compute Cost"}
+              </Button>
+            </Col>
+          </Row>
+        </Modal.Body>
+      </Modal>
+      <Container fluid={true} className='dasbwr'>
         <Helmet>
-          <meta charSet="utf-8" />
+          <meta charSet='utf-8' />
           <title>Molecular - Admin Work Order</title>
           <link />
         </Helmet>
         <Row>
           <DashboardNav />
-          <div id="overview"></div>
+          <div id='overview'></div>
         </Row>
-        <Row className="rowt3 row3t2">
-          <Col md={12} className="job34">
-            <div className="title_wo title_wo12 title_wo_">
-              <div className="workorderheader fixedtitle">
-                <Link to="/admin_work_order">
+        <Row className='rowt3 row3t2'>
+          <Col md={12} className='job34'>
+            <div className='title_wo title_wo12 title_wo_'>
+              <div className='workorderheader fixedtitle'>
+                <span
+                  onClick={() => window.history.back()}
+                  className='curspointer'>
                   {" "}
-                  <img src={arrowback} className="arrowback" />
-                </Link>
+                  <img src={arrowback} className='arrowback' />
+                </span>
                 Work Details
               </div>
             </div>
             {new_work && (
-              <div className="rjwrapper mrgin__right">
+              <div className='rjwrapper mrgin__right'>
                 {/* <Button
                   className="accjct1"
                   onClick={() => {
@@ -988,102 +1512,133 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                 </Button> */}
               </div>
             )}
-            {inreview &&
-              work_order_detail?.actions?.canAssignSpecialist == false && (
-                <div className="raise1">
-                  <div className="rjwrapper mrgin__right">
-                    <Link to="/work_order_evaluation">
-                      <Button className=" raise_inv">{"Edit"}</Button>
-                    </Link>
+            <div className='fllx2'>
+              {inreview &&
+                work_order_detail?.actions?.canAssignSpecialist == false && (
+                  <div className='raise1'>
+                    <div className='rjwrapper mrgin__right'>
+                      <Link to='/work_order_evaluation'>
+                        <Button className=' raise_inv'>{"Edit"}</Button>
+                      </Link>
+                    </div>
                   </div>
-                </div>
-              )}
-            {inreview && work_order_detail?.assigned_specialists.length !== 0 && (
-              <div className="raise1">
-                <div className="rjwrapper mrgin__right">
-                  <Button
-                    className=" raise_inv startproject"
-                    onClick={StartProject}
-                  >
-                    {!isloading ? "Start Work Order" : "Starting"} <Play />
-                  </Button>
-                </div>
-              </div>
-            )}
-            {work_order_detail?.status == "Active" &&
-              work_order_detail?.assigned_specialists?.length !== 0 && (
-                <div className="raise1">
-                  <div className="rjwrapper mrgin__right">
-                    <Button
-                      className=" raise_inv startproject"
-                      onClick={StopProject}
-                    >
-                      {!isloading ? "End Work Order" : "Processing"} <Stop />
+                )}
+              {inreview && work_order_detail?.actions?.canCost == true && (
+                <div className='raise1'>
+                  <div className='rjwrapper mrgin__right'>
+                    <Button className=' raise_inv' onClick={modalShow1}>
+                      {"Cost Settings"}
                     </Button>
                   </div>
                 </div>
               )}
-            <Row className="mgtop">
-              <Col md={2} className="job23_ mheight_">
-                <p className="exp23">
-                  <img src={portfolio} alt="portfolio" className="portfolioq" />
+              {inreview &&
+                work_order_detail?.assigned_specialists.length !== 0 && (
+                  <div className='raise1'>
+                    <div className='rjwrapper mrgin__right'>
+                      <Button
+                        className=' raise_inv startproject'
+                        onClick={StartProject}>
+                        {!isloading ? "Commence Work" : "Processing"} <Play />
+                      </Button>
+                    </div>
+                  </div>
+                )}
+              {work_order_detail?.status == "Active" &&
+                work_order_detail?.assigned_specialists?.length !== 0 && (
+                  <div className='raise1'>
+                    <div className='rjwrapper mrgin__right'>
+                      <Button
+                        className=' raise_inv startproject'
+                        onClick={StopProject}>
+                        {!isloading ? "End Work" : "Processing"} <Stop />
+                      </Button>
+                    </div>
+                  </div>
+                )}
+              {work_order_detail?.actions?.canSendSla && (
+                <div className='raise1'>
+                  <div className='rjwrapper mrgin__right'>
+                    {!work_order_detail?.sla_sent && (
+                      <Button
+                        className=' raise_inv startproject'
+                        onClick={sendSLA}>
+                        {!isloading ? "Send SLA" : "Processing"}
+                      </Button>
+                    )}
+                    {work_order_detail?.sla_sent && (
+                      <Button
+                        className=' raise_inv startproject'
+                        onClick={sendSLA}>
+                        {!isloading ? "Resend SLA" : "Processing"}
+                      </Button>
+                    )}
+                  </div>
+                </div>
+              )}
+            </div>
+
+            <Row className='mgtop'>
+              <Col md={2} className='job23_ mheight_'>
+                <p className='exp23'>
+                  <img src={portfolio} alt='portfolio' className='portfolioq' />
                 </p>
 
-                <p className="bview">
-                  <NavHashLink to="#overview">Overview</NavHashLink>
+                <p className='bview'>
+                  <NavHashLink to='#overview'>Overview</NavHashLink>
                 </p>
-                <p className="bview inactive_bv">
+                <p className='bview inactive_bv'>
                   <NavHashLink
-                    className="details"
-                    to="#details"
+                    className='details'
+                    to='#details'
                     // activeStyle={{ background: "#fd8b003b", color: "#fd8c00" }}
                   >
                     Specialist Details
                   </NavHashLink>
                 </p>
-                <p className="bview inactive_bv">
-                  <NavHashLink to="#work">Work Details</NavHashLink>
+                <p className='bview inactive_bv'>
+                  <NavHashLink to='#work'>Work Details</NavHashLink>
                 </p>
-                <p className="bview inactive_bv">
-                  <NavHashLink to="#worksheet">Work Sheet</NavHashLink>
+                <p className='bview inactive_bv'>
+                  <NavHashLink to='#worksheet'>Work Sheet</NavHashLink>
                 </p>
-                <p className="bview inactive_bv">
-                  <NavHashLink to="#actioninvoice">Invoice</NavHashLink>
+                <p className='bview inactive_bv'>
+                  <NavHashLink to='#actioninvoice'>Invoice</NavHashLink>
                 </p>
 
                 {/* <p className="bview inactive_bv">
                   <a href="#actions">Actions</a>
                 </p> */}
               </Col>
-              <Col md={10} className="job23_1a_ job23_1a_p">
-                <div className="job23_1a">
-                  <div className="">
+              <Col md={10} className='job23_1a_ job23_1a_p'>
+                <div className='job23_1a'>
+                  <div className=''>
                     <WorkOrderCardsMinInfo order_detail={work_order_detail} />
                   </div>
                 </div>
-                <div className="job23_1a" id="details">
+                <div className='job23_1a' id='details'>
                   {assigned_specialists.length == 0 &&
                     !new_work &&
                     work_order_detail.invoice && (
                       <>
-                        <h6 className="title22">Specialists Invitation</h6>
-                        <Col md={11} className="containerforemptyorder1 cust20">
-                          <div className="containerforemptyorder">
+                        <h6 className='title22'>Specialists Invitation</h6>
+                        <Col md={11} className='containerforemptyorder1 cust20'>
+                          <div className='containerforemptyorder'>
                             <img
                               src={no_work_order}
                               alt={"no_work_order"}
-                              className="no_work_order"
+                              className='no_work_order'
                             />
                           </div>
-                          <div className="no_work1">
+                          <div className='no_work1'>
                             <div>No Specialist have been invited</div>
                             {work_order_detail?.invoice?.approved == null &&
                               "Awaiting invoice approval"}
                           </div>
                           {work_order_detail?.invoice?.approved && (
-                            <div className="nojob2 ">
+                            <div className='nojob2 '>
                               <div
-                                className="job3 job_1"
+                                className='job3 job_1'
                                 onClick={() => {
                                   localStorage.setItem(
                                     "work_order_details",
@@ -1092,8 +1647,7 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                                   props.history.push(
                                     "/admin_assign_specialist"
                                   );
-                                }}
-                              >
+                                }}>
                                 Invite specialist
                               </div>
                             </div>
@@ -1103,40 +1657,39 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                     )}
                   {work_order_detail.invoice == null &&
                     work_order_detail.status !== "New" && (
-                      <Col md={11} className="containerforemptyorder1 cust20">
-                        <div className="containerforemptyorder">
+                      <Col md={11} className='containerforemptyorder1 cust20'>
+                        <div className='containerforemptyorder'>
                           <img
                             src={no_work_order}
                             alt={"no_work_order"}
-                            className="no_work_order"
+                            className='no_work_order'
                           />
                         </div>
-                        <div className="no_work1">
+                        <div className='no_work1'>
                           Proforma Invoice has not been raised
                         </div>
-                        <div className="nojob2 ">
+                        <div className='nojob2 '>
                           <div
-                            className="job3 job_1"
+                            className='job3 job_1'
                             onClick={() => {
                               props.history.push("/raise_proforma_invoice");
-                            }}
-                          >
+                            }}>
                             Raise Proforma Invoice
                           </div>
                         </div>
                       </Col>
                     )}
-                  <div className="dplsplsacc">
+                  <div className='dplsplsacc'>
                     <Card>
                       <div onClick={toggleAccordion}>
-                        <div className="deploydsplstwrapp">
+                        <div className='deploydsplstwrapp'>
                           <div>
-                            <span className="deplyeaggrgt">
+                            <span className='deplyeaggrgt'>
                               {props?.group_data?.total_members} Invited
                               Specialist
                             </span>
                           </div>
-                          <div className="accimgwrap">
+                          <div className='accimgwrap'>
                             <span>
                               <img
                                 src={chevrondown}
@@ -1147,42 +1700,41 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                         </div>
                       </div>
                     </Card>
-                    {isloading && <Spinner animation={"grow"} variant="info" />}
+                    {isloading && <Spinner animation={"grow"} variant='info' />}
                     <div
                       style={{ maxHeight: `${collapseHeight}` }}
-                      className="acccollapsediv"
-                      ref={content}
-                    >
+                      className='acccollapsediv maxh-378'
+                      ref={content}>
                       {!new_work && assigned_specialists.length !== 0 && (
                         <>
-                          <div className="group_flex">
+                          <div className='group_flex'>
                             {/* <div className="grpA">
                             Group <b>A</b>
                           </div> */}
-                            <div className="grpB">
+                            <div className='grpB'>
                               <b>
                                 {work_order_detail?.total_assigned_specialists}
                               </b>{" "}
                               Invited
                             </div>
                             <div
-                              className="job3 job_1 job_12"
+                              className='job3 job_1 job_12'
                               onClick={() => {
                                 localStorage.setItem(
                                   "work_order_details",
                                   JSON.stringify(work_order_detail)
                                 );
                                 props.history.push("/admin_assign_specialist");
-                              }}
-                            >
+                              }}>
                               Invite specialist
                             </div>
                           </div>
-                          <div className="tabledata tabledataweb">
-                            <div className="header_12 pleft">Fullname</div>
-                            <div className="header_12">Type</div>
-                            <div className="header_12">Group Position</div>
-                            <div className="header_12">Status</div>
+                          <div className='tabledata tabledataweb'>
+                            <div className='header_12 pleft'>Fullname</div>
+                            <div className='header_12'>Type</div>
+                            <div className='header_12'>Spread Position</div>
+                            <div className='header_12'>Status</div>
+                            <div className='header_12'>Action</div>
                           </div>
                           {assigned_specialists.length !== 0 &&
                             assigned_specialists
@@ -1194,21 +1746,20 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                                       checkIfIsOdd(i)
                                         ? "tabledata"
                                         : "tabledata tablecontent"
-                                    }
-                                  >
-                                    <div className="header_12">
+                                    }>
+                                    <div className='header_12'>
                                       <img
                                         src={avatar_test}
-                                        className="specialist_avatar"
+                                        className='specialist_avatar'
                                       />
-                                      <div className="mobiletabledata">
+                                      <div className='mobiletabledata'>
                                         Fullname
                                       </div>
                                       {data.first_name}
                                       {data.last_name}
                                     </div>
-                                    <div className="header_12 typ22">
-                                      <div className="mobiletabledata mobiletabledata22">
+                                    <div className='header_12 typ22'>
+                                      <div className='mobiletabledata mobiletabledata22'>
                                         Type
                                       </div>
                                       <div>
@@ -1216,20 +1767,20 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                                         {capitalize(data.skills?.[0].name)}
                                       </div>
                                     </div>
-                                    <div className="header_12">
-                                      <div className="mobiletabledata mobiletabledata22">
-                                        Group Position
+                                    <div className='header_12'>
+                                      <div className='mobiletabledata mobiletabledata22'>
+                                        Spread Position
                                       </div>
-                                      <div className="glead"> Member </div>
+                                      <div className='glead'> Member </div>
                                     </div>
-                                    <div className="header_12 active_member">
-                                      <div className="mobiletabledata mobiletabledata22">
+                                    <div className='header_12 active_member'>
+                                      <div className='mobiletabledata mobiletabledata22'>
                                         Status
                                       </div>
-                                      <div className="active_member">
+                                      <div className='active_member'>
                                         {" "}
                                         {data.status == "Pending" ? (
-                                          <span className="pending_color">
+                                          <span className='pending_color'>
                                             {data.status}
                                           </span>
                                         ) : (
@@ -1237,14 +1788,23 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                                         )}
                                       </div>
                                     </div>
+                                    <div className='member'>
+                                      {data.status == "Pending" && (
+                                        <div
+                                          className='job3 job_1 job_12 text-white specailist_rv'
+                                          onClick={() => modalShow2(data.id)}>
+                                          Remove Specialist
+                                        </div>
+                                      )}
+                                    </div>
                                   </div>
                                 </>
                               ))}
-                          <div className="text-center">
+                          <div className='text-center'>
                             {" "}
-                            <span className="viewall_">
+                            <span className='viewall_'>
                               {" "}
-                              <Link to="/deployedspecialist">
+                              <Link to={`/deployedspecialist/${props.match.params.id}`}>
                                 View all
                               </Link>{" "}
                             </span>{" "}
@@ -1254,15 +1814,15 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
                     </div>
                   </div>
                   {/* invited specialist ends */}
-                  <div className="job23_1a wrap_z paddtop">
-                    <div className="active_member23">
-                      <div className="dplsplsacc">
+                  <div className='job23_1a wrap_z paddtop'>
+                    <div className='active_member23'>
+                      <div className='dplsplsacc'>
                         <Work_Details work_order_detailz={work_order_detail} />
                       </div>
                     </div>
                   </div>
                   <Work_Sheet />
-                  <div id="worksheet"></div>
+                  <div id='worksheet'></div>
                   <br />
                   {work_order_detail?.invoice?.length !== 0 && (
                     <>
@@ -1280,14 +1840,14 @@ const AdminViewWorkOrderDetails = withRouter((props: any) => {
       <ToastContainer
         enableMultiContainer
         containerId={"D"}
-        toastClassName="bg-danger text-white"
+        toastClassName='bg-danger text-white'
         hideProgressBar={true}
         position={"top-right"}
       />
       <ToastContainer
         enableMultiContainer
         containerId={"B"}
-        toastClassName="bg-orange text-white"
+        toastClassName='bg-orange text-white'
         hideProgressBar={true}
         position={"top-right"}
       />
