@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { Col, Row, Container, Form, Modal } from "react-bootstrap";
 import "../contractor.css";
 import DashboardNav from "../navbar";
@@ -17,6 +17,7 @@ import {
 import allCountries from "../../../listOfCountriesInTheWorld";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import camimg from "../../../images/imagecam.png";
 
 const Contractor_Profile = withRouter((props) => {
   const [state, setState] = useState({
@@ -29,14 +30,23 @@ const Contractor_Profile = withRouter((props) => {
     show: false,
     contractor: {},
     company_name: "",
+    contacts: [],
+    roles: [],
+    logo: "",
     website_url: "",
     sector: "",
     state_: "",
     address: "",
+    role: "",
     phone_number: "",
     email: "",
     first_name: "",
     last_name: "",
+    role1: "",
+    phone_number1: "",
+    email1: "",
+    first_name1: "",
+    last_name1: "",
     middle_name: "",
     reason: "",
     industry: "",
@@ -46,6 +56,19 @@ const Contractor_Profile = withRouter((props) => {
     confirm_password: "",
     list_of_industries: [],
     industry_id: "",
+    role2: "",
+    role_id1: "",
+    role_id2: "",
+    role_id3: "",
+    phone_number2: "",
+    email2: "",
+    first_name2: "",
+    last_name2: "",
+    role3: "",
+    phone_number3: "",
+    email3: "",
+    first_name3: "",
+    last_name3: "",
     isloading: false,
   });
   const onchange = (e) => {
@@ -155,12 +178,12 @@ const Contractor_Profile = withRouter((props) => {
         });
       });
   };
-  const CreateContactPerson = (role) => {
+  const CreateContactPerson = () => {
     const data = {
       first_name,
       last_name,
       email,
-      phone:phone_number,
+      phone: phone_number,
       role,
     };
     console.log(data);
@@ -181,17 +204,17 @@ const Contractor_Profile = withRouter((props) => {
         });
       })
       .catch((err) => {
-        if(err?.response?.status==406){
-          return notify(err?.response?.data?.errors?.email?.join(""),"D")
-        }
-        notify("Failed to create", "D");
         setState({
           ...state,
           isloading: false,
         });
+        if (err?.response?.status == 406) {
+          notify(err?.response?.data?.errors?.email?.join(""), "D");
+          return notify(err?.response?.data?.errors?.role?.join(""), "D");
+        }
+        notify("Failed to create", "D");
       });
   };
-
   const SubmitPassword = () => {
     const data = {
       old_password: current_password,
@@ -223,12 +246,11 @@ const Contractor_Profile = withRouter((props) => {
         });
       });
   };
-  
   useEffect(() => {
     window.scrollTo(0, 0);
     const token = contractorToken();
     if (token.user_type !== "contractor") {
-      return props.history.push("/login");
+      return props.history.push("/signin");
     }
     axios
       .all([
@@ -238,17 +260,39 @@ const Contractor_Profile = withRouter((props) => {
         axios.get(`${API}/industries`, {
           headers: { Authorization: `Bearer ${token.access_token}` },
         }),
+        axios.get(`${API}/contractor/contact/roles`, {
+          headers: { Authorization: `Bearer ${token.access_token}` },
+        }),
       ])
       .then(
-        axios.spread((res, res2) => {
+        axios.spread((res, res2, res3) => {
           console.log(res.data);
           setState({
             ...state,
             ...res.data.data,
             list_of_industries: res2.data.data,
+            roles: res3.data.data,
             contractor: res.data.data,
             state_: res.data.data.state,
             industry_id: res.data.data.industry_id,
+            role1: res?.data?.data?.contacts?.[0]?.role,
+            phone_number1: res?.data?.data?.contacts?.[0]?.phone,
+            email1: res?.data?.data?.contacts?.[0]?.email,
+            first_name1: res?.data?.data?.contacts?.[0]?.first_name,
+            last_name1: res?.data?.data?.contacts?.[0]?.last_name,
+            role2: res?.data?.data?.contacts?.[1]?.role,
+            phone_number2: res?.data?.data?.contacts?.[1]?.phone,
+            email2: res?.data?.data?.contacts?.[1]?.email,
+            first_name2: res?.data?.data?.contacts?.[1]?.first_name,
+            last_name2: res?.data?.data?.contacts?.[1]?.last_name,
+            role3: res?.data?.data?.contacts?.[2]?.role,
+            phone_number3: res?.data?.data?.contacts?.[2]?.phone,
+            email3: res?.data?.data?.contacts?.[2]?.email,
+            first_name3: res?.data?.data?.contacts?.[2]?.first_name,
+            last_name3: res?.data?.data?.contacts?.[2]?.last_name,
+            role_id1: res?.data?.data?.contacts?.[0]?.role_id,
+            role_id2: res?.data?.data?.contacts?.[1]?.role_id,
+            role_id3: res?.data?.data?.contacts?.[2]?.role_id,
           });
         })
       )
@@ -256,7 +300,94 @@ const Contractor_Profile = withRouter((props) => {
         console.log(err);
       });
   }, []);
-
+  const hiddenFileInput: any = useRef();
+  const handleClick = (event) => {
+    hiddenFileInput.current.click();
+  };
+  const handleImageChange = (e) => {
+    const reader: any = new FileReader();
+    reader.onload = () => {
+      if (reader.readyState === 2) {
+        setState({
+          ...state,
+          logo: reader.result,
+        });
+      }
+    };
+    reader.readAsDataURL(e.target.files[0]);
+    console.log(logo);
+    console.log(e.target.files[0]);
+    // upload image to server;
+    const availableToken = localStorage.getItem("loggedInDetails");
+    console.log(availableToken);
+    const token = availableToken ? JSON.parse(availableToken) : "";
+    console.log(token);
+    const imageData = new FormData();
+    imageData.append("logo", e.target.files[0]);
+    console.log(imageData);
+    axios
+      .post(`${API}/contractor/logo`, imageData, {
+        headers: {
+          Authorization: `Bearer ${token.access_token}`,
+          Accept: "application/json",
+          "Content-Type": "multipart/form-data",
+        },
+      })
+      .then((res) => {
+        console.log(res.data);
+        setTimeout(() => {
+          notify("image uploaded Successfully");
+        }, 1000);
+      })
+      .catch((err) => {
+        console.log(err.response);
+        notify("failed to Upload Image");
+      });
+  };
+  const UpdateContactPerson = (
+    id: any,
+    first_name,
+    last_name,
+    email,
+    phone,
+    role
+  ) => {
+    const data = {
+      first_name,
+      last_name,
+      email,
+      phone,
+      role,
+    };
+    console.log(data);
+    setState({
+      ...state,
+      isloading: true,
+    });
+    axios
+      .put(`${API}/contractor/contacts/${id}`, data, {
+        headers: { Authorization: `Bearer ${contractorToken().access_token}` },
+      })
+      .then((res) => {
+        console.log(res);
+        notify("Updated successfully");
+        setState({
+          ...state,
+          isloading: false,
+        });
+      })
+      .catch((err) => {
+        setState({
+          ...state,
+          isloading: false,
+        });
+        if (err?.response?.status == 406) {
+          notify(err?.response?.data?.errors?.email?.join(""), "D");
+          return notify(err?.response?.data?.errors?.role?.join(""), "D");
+        }
+        notify("Failed to update", "D");
+      });
+  };
   const {
     firsttab,
     secondtab,
@@ -265,8 +396,11 @@ const Contractor_Profile = withRouter((props) => {
     show,
     company_name,
     website_url,
-    sector,
+    contacts,
+    role,
+    logo,
     country,
+    roles,
     old_password,
     state_,
     address,
@@ -282,7 +416,26 @@ const Contractor_Profile = withRouter((props) => {
     current_password,
     new_password,
     confirm_password,
-  } = state;
+    role1,
+    phone_number1,
+    email1,
+    first_name1,
+    last_name1,
+    role2,
+    phone_number2,
+    email2,
+    first_name2,
+    last_name2,
+    role3,
+    phone_number3,
+    email3,
+    first_name3,
+    last_name3,
+    role_id1,
+    role_id2,
+    role_id3,
+  }: any = state;
+  console.log(contacts);
   return (
     <>
       <Container fluid={true} className="dasbwr">
@@ -304,10 +457,36 @@ const Contractor_Profile = withRouter((props) => {
                 <div className="titleprofile1">Settings</div>
                 <div className="setting1">
                   <div className="namestyle11">
-                    <span className="namestyle">
-                      {capitalize(company_name?.split("")[0])}
-                      {capitalize(company_name?.split("")[1])}
-                    </span>
+                    <div>
+                      <span className="spluserimg">
+                        {logo ? (
+                          <img src={logo} className="useravatar" />
+                        ) : (
+                          <span className="namestyle">
+                            {capitalize(company_name?.split("")[0])}
+                            {capitalize(company_name?.split("")[1])}
+                          </span>
+                        )}
+                      </span>
+                      <div className="camdv">
+                        <img
+                          src={camimg}
+                          className="user-cam-img"
+                          alt="cam-img"
+                          onClick={handleClick}
+                        />
+                      </div>
+                      <input
+                        type="file"
+                        onChange={handleImageChange}
+                        style={{ display: "none" }}
+                        accept="image/*"
+                        ref={hiddenFileInput}
+                      />
+                      <p className="upldtxt" onClick={handleClick}>
+                        Upload Picture
+                      </p>
+                    </div>
                     <div className="home_pone12">
                       <div className="helmot">{company_name}</div>
                       <div className="helmot11">{industry}</div>
@@ -497,226 +676,422 @@ const Contractor_Profile = withRouter((props) => {
                   {/* Second Tab Starts*/}
                   {secondtab && (
                     <>
-                      <Row className="section_form1">
-                        <Col md={12}>
-                          <h3 className="userprofile userprofile12 boldtext">
-                            Contact Person 1
-                          </h3>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              First Name
-                            </h6>
-                            <Form.Control
-                              className="userfield"
-                              id="first_name"
-                              value={first_name}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              Last Name
-                            </h6>
-                            <Form.Control
-                              type="text"
-                              className="userfield"
-                              id="last_name"
-                              value={last_name}
-                              onChange={onchange}
-                              placeholder=""
-                            />
-                          </Form.Group>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">Email</h6>
-                            <Form.Control
-                              type="text"
-                              className="userfield"
-                              id="email"
-                              value={email}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                      </Row>
-                      <Row>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              Phone Number
-                            </h6>
-                            <Form.Control
-                              className="userfield"
-                              id="phone_number"
-                              value={phone_number}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                      </Row>
-                      <Row>
-                          <Col md={12}>
-                            <div
-                              className="job31"
-                              onClick={()=>CreateContactPerson(1)}
-                            >
-                              {isloading ? "Creating" : "Create"}
-                            </div>
-                          </Col>
-                        </Row>
-                      { false && 
-                      <>
-                      <Row className="section_form1">
-                        <Col md={12}>
-                          <h3 className="userprofile userprofile12 boldtext">
-                            Contact Person 2
-                          </h3>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              First Name
-                            </h6>
-                            <Form.Control
-                              className="userfield"
-                              id="first_name"
-                              value={first_name}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              Last Name
-                            </h6>
-                            <Form.Control
-                              type="text"
-                              className="userfield"
-                              id="last_name"
-                              value={last_name}
-                              onChange={onchange}
-                              placeholder=""
-                            />
-                          </Form.Group>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">Email</h6>
-                            <Form.Control
-                              type="text"
-                              className="userfield"
-                              id="email"
-                              value={email}
-                              onChange={onInputChange}
-                            />
-                          </Form.Group>
-                        </Col>
-                      </Row>
-                      <Row>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              Phone Number
-                            </h6>
-                            <Form.Control
-                              className="userfield"
-                              id="phone_number"
-                              value={phone_number}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                      </Row>
-                      <Row>
-                          <Col md={12}>
-                            <div
-                              className="job31"
-                              onClick={()=>CreateContactPerson(2)}
-                            >
-                              {isloading ? "Creating" : "Create"}
-                            </div>
-                          </Col>
-                        </Row>
-                      <Row className="section_form1">
-                        <Col md={12}>
-                          <h3 className="userprofile userprofile12 boldtext">
-                            Contact Person 3
-                          </h3>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              First Name
-                            </h6>
-                            <Form.Control
-                              className="userfield"
-                              id="first_name"
-                              value={first_name}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              Last Name
-                            </h6>
-                            <Form.Control
-                              type="text"
-                              className="userfield"
-                              id="last_name"
-                              value={last_name}
-                              onChange={onchange}
-                              placeholder=""
-                            />
-                          </Form.Group>
-                        </Col>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">Email</h6>
-                            <Form.Control
-                              type="text"
-                              className="userfield"
-                              id="email"
-                              value={email}
-                              onChange={onInputChange}
-                            />
-                          </Form.Group>
-                        </Col>
-                      </Row>
-                      <Row>
-                        <Col md={4} className="formsection1">
-                          <Form.Group>
-                            <h6 className="userprofile userprofile12">
-                              Phone Number
-                            </h6>
-                            <Form.Control
-                              className="userfield"
-                              id="phone_number"
-                              value={phone_number}
-                              onChange={onchange}
-                            />
-                          </Form.Group>
-                        </Col>
-                      </Row>
-                      <Row>
-                          <Col md={12}>
-                            <div
-                              className="job31"
-                              onClick={()=>CreateContactPerson(3)}
-                            >
-                              {isloading ? "Creating" : "Create"}
-                            </div>
-                          </Col>
-                        </Row>
-                      </>
-                      }
+                      {contacts?.length < 3 && (
+                        <>
+                          <Row className="section_form1">
+                            <Col md={12}>
+                              <h3 className="userprofile userprofile12 boldtext">
+                                Create Contact Person ({" "}
+                                <small>maximum of 3</small> )
+                              </h3>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  First Name
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="first_name"
+                                  value={first_name}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Last Name
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="last_name"
+                                  value={last_name}
+                                  onChange={onchange}
+                                  placeholder=""
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Email
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="email"
+                                  value={email}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Phone Number
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="phone_number"
+                                  value={phone_number}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Roles
+                                </h6>
+                                <select
+                                  name="role"
+                                  id="role"
+                                  onChange={onchange}
+                                  className="form-control userfield"
+                                >
+                                  <option>{role ? role : ""}</option>
+                                  {roles?.map((data: any, i) => (
+                                    <option key={i} value={data?.id}>
+                                      {data?.name}
+                                    </option>
+                                  ))}
+                                </select>
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={12}>
+                              <div
+                                className="job31"
+                                onClick={() => CreateContactPerson()}
+                              >
+                                {isloading ? "Creating" : "Create"}
+                              </div>
+                            </Col>
+                          </Row>
+                        </>
+                      )}
+                      {contacts[0] && (
+                        <>
+                          <Row className="section_form1">
+                            <Col md={12}>
+                              <h3 className="userprofile userprofile12 boldtext">
+                                Update Contact Person
+                              </h3>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  First Name
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="first_name1"
+                                  value={first_name1}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Last Name
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="last_name1"
+                                  value={last_name1}
+                                  onChange={onchange}
+                                  placeholder=""
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Email
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="email1"
+                                  value={email1}
+                                  onChange={onInputChange}
+                                />
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Phone Number
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="phone_number1"
+                                  value={phone_number1}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Role
+                                </h6>
+                                <select
+                                  name="role_id1"
+                                  id="role_id1"
+                                  onChange={onchange}
+                                  className="form-control userfield"
+                                >
+                                  <option>{role1 ? role1 : ""}</option>
+                                  {roles?.map((data: any, i) => (
+                                    <option key={i} value={data?.id}>
+                                      {data?.name}
+                                    </option>
+                                  ))}
+                                </select>
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={12}>
+                              <div
+                                className="job31"
+                                onClick={() =>
+                                  UpdateContactPerson(
+                                    contacts[0]?.id,
+                                    first_name1,
+                                    last_name1,
+                                    email1,
+                                    phone_number1,
+                                    role_id1
+                                  )
+                                }
+                              >
+                                {isloading ? "Updating" : "Update"}
+                              </div>
+                            </Col>
+                          </Row>
+                        </>
+                      )}
+                      {contacts[1] && (
+                        <>
+                          <Row className="section_form1">
+                            <Col md={12}>
+                              <h3 className="userprofile userprofile12 boldtext"></h3>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  First Name
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="first_name2"
+                                  value={first_name2}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Last Name
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="last_name2"
+                                  value={last_name2}
+                                  onChange={onchange}
+                                  placeholder=""
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Email
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="email2"
+                                  value={email2}
+                                  onChange={onInputChange}
+                                />
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Phone Number
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="phone_number2"
+                                  value={phone_number2}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Role
+                                </h6>
+                                <select
+                                  name="role_id2"
+                                  id="role_id2"
+                                  onChange={onchange}
+                                  className="form-control userfield"
+                                >
+                                  <option>{role2 ? role2 : ""}</option>
+                                  {roles?.map((data: any, i) => (
+                                    <option key={i} value={data?.id}>
+                                      {data?.name}
+                                    </option>
+                                  ))}
+                                </select>
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={12}>
+                              <div
+                                className="job31"
+                                onClick={() =>
+                                  UpdateContactPerson(
+                                    contacts[1]?.id,
+                                    first_name2,
+                                    last_name2,
+                                    email2,
+                                    phone_number2,
+                                    role_id2
+                                  )
+                                }
+                              >
+                                {isloading ? "Updating" : "Update"}
+                              </div>
+                            </Col>
+                          </Row>
+                        </>
+                      )}
+                      {contacts[2] && (
+                        <>
+                          <Row className="section_form1">
+                            <Col md={12}>
+                              <h3 className="userprofile userprofile12 boldtext"></h3>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  First Name
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="first_name3"
+                                  value={first_name3}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Last Name
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="last_name3"
+                                  value={last_name3}
+                                  onChange={onchange}
+                                  placeholder=""
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Email
+                                </h6>
+                                <Form.Control
+                                  type="text"
+                                  className="userfield"
+                                  id="email3"
+                                  value={email3}
+                                  onChange={onInputChange}
+                                />
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Phone Number
+                                </h6>
+                                <Form.Control
+                                  className="userfield"
+                                  id="phone_number3"
+                                  value={phone_number3}
+                                  onChange={onchange}
+                                />
+                              </Form.Group>
+                            </Col>
+                            <Col md={4} className="formsection1">
+                              <Form.Group>
+                                <h6 className="userprofile userprofile12">
+                                  Role
+                                </h6>
+                                <select
+                                  name="role_id3"
+                                  id="role_id3"
+                                  onChange={onchange}
+                                  className="form-control userfield"
+                                >
+                                  <option>{role3 ? role3 : ""}</option>
+                                  {roles?.map((data: any, i) => (
+                                    <option key={i} value={data?.id}>
+                                      {data?.name}
+                                    </option>
+                                  ))}
+                                </select>
+                              </Form.Group>
+                            </Col>
+                          </Row>
+                          <Row>
+                            <Col md={12}>
+                              <div
+                                className="job31"
+                                onClick={() =>
+                                  UpdateContactPerson(
+                                    contacts[2]?.id,
+                                    first_name3,
+                                    last_name3,
+                                    email3,
+                                    phone_number3,
+                                    role_id3
+                                  )
+                                }
+                              >
+                                {isloading ? "Updating" : "Update"}
+                              </div>
+                            </Col>
+                          </Row>
+                        </>
+                      )}
                     </>
                   )}
                   {/* Second Tab ends*/}
