@@ -86,7 +86,8 @@ const Specialist_Payment_Invoice = (props) => {
       ...state,
       cycle_id: id,
       max_requested_amount: amount,
-      requested_amount: (rate * amount)?.toFixed(2),
+      requested_amount: 0.00,
+      // requested_amount: (rate * amount)?.toFixed(2),
       terminateWorkModal: true,
     });
   };
@@ -222,12 +223,12 @@ const fieldRef: any = useRef();
       )}
      <div className="splinvoicemodalmssgwrap">
        <i className="fa fa-exclamation fa-rotate-180 invoiceexclm" aria-hidden="true"></i>
-       <p>You can only make a maximum of  {current_currency}{ FormatAmount((0.65 * max_requested_amount)?.toFixed(2))}  from this cycle. </p>
+       <p>The amount of early payment is limited to {current_currency}{ FormatAmount((0.65 * max_requested_amount)?.toFixed(2))}. </p>
      </div>
-     {/* <div className="splinvoicemodalmssgwrap">
+     <div className="splinvoicemodalmssgwrap">
        <i className="fa fa-exclamation fa-rotate-180 invoiceexclm" aria-hidden="true"></i>
-       <p>Early payments attracts 5% charge of your amount from this cycle. </p>
-     </div>  */}
+       <p>Early payment attracts 5% charge. </p>
+     </div>
       <form>
         <Row>
            <Col md={12} className="formsection1">
@@ -238,12 +239,12 @@ const fieldRef: any = useRef();
                 <Form.Control
                   type="number"
                   name="requested_amount"
-                  value={requested_amount} 
+                  // value={requested_amount} 
                   className="userfield"
                   onChange={onchange}
                   placeholder="Amount"
                   min={0}
-                  max={0.65 * requested_amount} 
+                  max={(0.65 * requested_amount)?.toFixed(2)} 
                   />
             </Form.Group>
           </Col>
@@ -295,7 +296,7 @@ const fieldRef: any = useRef();
                   <p className="brkdwn-id">{invoice_details?.number}</p>
                 </div>
                 <div>
-                  <img src={logo} alt="molecular-logo" />
+                  <img src={logo} alt="molecular-logo" className="img-fluid" />
                 </div>
               </div>
               <p className="spltpaybreakdwn-title">
@@ -320,15 +321,15 @@ const fieldRef: any = useRef();
                 </div>
               </div>
               <div>
-                <Table hover>
+                <Table hover size="md" responsive>
                   <thead className="splinvoitablehead">
                     <tr>
                       <th>Cycle</th>
-                      <th>Payment Reference</th>
-                      <th>Amount</th>
-                      <th>Amount Paid</th>
-                      <th>Status</th>
-                      <th>Date</th>
+                      <th style={{minWidth: "10rem"}}>Payment Reference</th>
+                      <th style={{minWidth: "8rem"}}>Amount</th>
+                      <th style={{minWidth: "8rem"}}>Amount Paid</th>
+                      <th style={{minWidth: "9rem"}}>Status</th>
+                      <th style={{minWidth: "8rem"}}>Date</th>
                       <th></th>
                     </tr>
                   </thead>
@@ -349,21 +350,21 @@ const fieldRef: any = useRef();
                    </td>
                    <td>
                      {data.status == "Paid" && (
-                       <div className="invpaystatwrap">
-                       <span className="paystatindcator"></span>
-                       <span className="paystattext">paid</span>
+                       <div className="invpaystatwrap po912">
+                       <span className="paystatindcator po912"></span>
+                       <span className="paystattext po912">paid</span>
                      </div>
                     )}
                     {data.status == "Not Due" &&(
-                      <div className="invpaystatwrap pendinwrap">
-                      <span className="paystatindcator pendininvoice"></span>
-                      <span className="paystattext pendininvtext">Not Due</span>
+                      <div className="invpaystatwrap pendinwrap po912">
+                      <span className="paystatindcator pendininvoice po912"></span>
+                      <span className="paystattext pendininvtext po912">Not Due</span>
                     </div>
                     )}
                     {data.status == "Outstanding" &&(
-                      <div className="invpaystatwrap pendinwrap">
-                      <span className="paystatindcator pendininvoice"></span>
-                      <span className="paystattext pendininvtext">Outstanding</span>
+                      <div className="invpaystatwrap pendinwrap po912">
+                      <span className="paystatindcator pendininvoice po912"></span>
+                      <span className="paystattext pendininvtext po912">Outstanding</span>
                     </div>
                     )}
                    </td> 
@@ -391,13 +392,13 @@ const fieldRef: any = useRef();
             </div>
             <div className="spltpaybreakdwnwrapper">
               <div>
-                <Table hover>
+                <Table hover size="md" responsive>
                   <thead className="splinvoitablehead">
                     <tr>
                       <th>S/N</th>
-                      <th> Amount</th>
-                      <th>Status</th>
-                      <th>Date</th>
+                      <th style={{minWidth: "8rem"}}> Amount</th>
+                      <th style={{minWidth: "8rem"}}>Status</th>
+                      <th style={{minWidth: "8rem"}}>Date</th>
                     </tr>
                   </thead>
                   <tbody>
