@@ -6,7 +6,7 @@ import styled from "styled-components";
 
 const Ul = styled.ul`
      display: flex !important;
-     width: 55%;
+     width: 78%;
      color: #999999;
      font-size: 16px;
      font-weight: 600;
@@ -24,10 +24,8 @@ const Ul = styled.ul`
     a:hover{
       text-decoration: none;
   }
+
 @media screen and (max-width:1024px){
-     width: 73%;
-}
-@media screen and (max-width:769px){
     position: fixed;
     left: -1px;
     height: 92vh;
@@ -47,27 +45,49 @@ const Ul = styled.ul`
 const Navlist = ({ open }) => {
   return (
     <Ul open={open}>
-      <NavHashLink
+      {/* <NavHashLink
         to="/#home"
         className="homenavlnks"
         activeStyle={{
           color: "#222073",
           lineHeight: "29px",
-          borderBottom: "4px solid #fd8c00"
+          borderBottom: "4px solid #fd8c00",
         }}
       >
         <li>Home</li>
-      </NavHashLink>
-      <NavHashLink
+      </NavHashLink> */}
+      {/* <NavHashLink
         to="/#our_services"
         className="homenavlnks"
         activeStyle={{
           color: "#222073",
           lineHeight: "29px",
-          borderBottom: "4px solid #fd8c00"
+          borderBottom: "4px solid #fd8c00",
         }}
       >
         <li>Our Services</li>
+      </NavHashLink> */}
+      <NavHashLink
+        to="/contractorlanding"
+        activeStyle={{
+          color: "#222073",
+          lineHeight: "29px",
+          borderBottom: "4px solid #fd8c00",
+        }}
+        className="homenavlnks"
+      >
+        <li>Hire Specialist</li>
+      </NavHashLink>
+      <NavHashLink
+        to="/specialistlanding"
+        activeStyle={{
+          color: "#222073",
+          lineHeight: "29px",
+          borderBottom: "4px solid #fd8c00",
+        }}
+        className="homenavlnks"
+      >
+        <li>Find work</li>
       </NavHashLink>
       <NavHashLink
         to="/Industry_segments"
@@ -75,7 +95,7 @@ const Navlist = ({ open }) => {
         activeStyle={{
           color: "#222073",
           lineHeight: "29px",
-          borderBottom: "4px solid #fd8c00"
+          borderBottom: "4px solid #fd8c00",
         }}
       >
         <li>Industry Segments</li>
@@ -86,25 +106,29 @@ const Navlist = ({ open }) => {
         activeStyle={{
           color: "#222073",
           lineHeight: "29px",
-          borderBottom: "4px solid #fd8c00"
+          borderBottom: "4px solid #fd8c00",
         }}
       >
-        <li>Careers</li>
+        <li>What we do</li>
       </NavHashLink>
       <NavHashLink
-        to="/contactus"
+        to="/blog"
+        className="homenavlnks"
         activeStyle={{
           color: "#222073",
           lineHeight: "29px",
-          borderBottom: "4px solid #fd8c00"
+          borderBottom: "4px solid #fd8c00",
         }}
-        className="homenavlnks"
       >
-        <li>Contact us</li>
+        <li>Blog</li>
       </NavHashLink>
-      <Link to="/signin">
+      <NavHashLink
+        className="details"
+        to="/signin"
+        // activeStyle={{ background: "#fd8b003b", color: "#fd8c00" }}
+      >
         <li className="nav-login-btn">Login</li>
-      </Link>
+      </NavHashLink>
     </Ul>
   );
 };
